Add activity to new file
