+++ src/mahoji/lib/abstracted_commands/cancelTaskCommand.ts
@@ -2,14 +2,11 @@ import type { ChatInputCommandInteraction } from 'discord.js';
 
 import { formatDuration, mentionCommand } from '@oldschoolgg/toolkit';
 import { Time } from 'e';
-import { Bank } from 'oldschooljs';
-import { refundChargeBank } from '../../../lib/degradeableItems';
 import { cancelTask } from '../../../lib/settings/settings';
 import { ChargeBank } from '../../../lib/structures/Bank';
 import type { ActivityTaskOptions, NexTaskOptions, RaidsOptions, RefundOptions } from '../../../lib/types/minions';
 import addSubTaskToActivityTask from '../../../lib/util/addSubTaskToActivityTask';
 import { handleMahojiConfirmation } from '../../../lib/util/handleMahojiConfirmation';
-import { handleTripFinish } from '../../../lib/util/handleTripFinish';
 import { getActivityOfUser } from '../../../lib/util/minionIsBusy';
 
 export async function cancelTaskCommand(
@@ -96,26 +93,3 @@ Please confirm if you want to call your minion back from their trip.${refundMess
 
 	return `${mName}'s trip was cancelled, and they're now available.`;
 }
-
-export const RefundTask: MinionTask = {
-	type: 'Refund',
-	async run(data: RefundOptions) {
-		const user = await mUserFetch(data.userID);
-		let str = `${user}, ${user.minionName}'s trip has been cancelled, and they're now available. `;
-
-		const items = new Bank(data.refundItems);
-		const charges = new ChargeBank(data.refundCharges);
-
-		if (data.refundItems && items.length > 0) {
-			await user.addItemsToBank({ items: data.refundItems });
-			str += `They have returned with the following supplies: ${items.toString()}. `;
-		}
-
-		if (data.refundCharges && charges.length() > 0) {
-			const results = await refundChargeBank(user, charges);
-			str += results.map(r => r.userMessage).join(' ');
-		}
-
-		return handleTripFinish(user, data.channelID, str, undefined, data, null);
-	}
-};
